requirements.txt
pydantic==2.6.3
pydantic-settings==2.2.1
python-telegram-bot[webhooks]==21.5
pymongo==4.4.0
pre-commit==3.7.1
